feat: 폴더 정보 버튼 구현
